Give Application a default type parameter `Canvas`, which is `WindowCanvas`, and ensure the `Engine` is passed an `&'a dyn mut Application<WindowCanvas>`
The idea here is to allow extensions to wrap an `impl Application` with
their own custom Canvas type, then implement `Application<WindowCanvas>` themselves. The extension can then be passed to the engine, but is free to control how the user's actions are actually drawn.

For example, a "Transformed View" extension can implement it's own
custom `Canvas` type that wraps a `WindowCanvas`, and allows the
viewport, zoom level, etc to be changed by the Application implementor.